tsql-core/src/test/java/com/xy/xsql/block/tsql/core/clause/select/IntoConverterTest.java
package com.xy.xsql.block.tsql.core.clause.select;

import com.xy.xsql.block.core.ReferenceBlockPrinter;
import com.xy.xsql.block.model.ReferenceBlock;
import org.junit.Assert;
import org.junit.Test;

import java.io.StringWriter;

import static org.junit.Assert.*;

/**
 * Created by xiaoyao9184 on 2017/6/21.
 */
public class IntoConverterTest {

    @Test
    public void test() throws Exception {
        ReferenceBlock b = IntoConverter.meta();

        StringWriter writer = new ReferenceBlockPrinter()
                .print(b);

        System.out.print(writer);
        Assert.assertEquals(writer.toString(),
                "<INTO Clause> ::=\n" +
                        "INTO new_table [ ON filegroup ]");
    }

}